fix average rating

+++ src/content/model/AmazonSearchItem.ts
@@ -32,7 +32,7 @@ export class AmazonSearchItem {
         try {
             const ratingEl = this.htmlElement.getElementsByClassName("a-section a-spacing-none a-spacing-top-micro")[0];
             let avgRatingTxt = ratingEl.getElementsByClassName("a-icon-alt")[0].textContent;
-            this.avgRating = parseInt(avgRatingTxt.slice(0, 3)) * 10;
+            this.avgRating = parseFloat(avgRatingTxt.slice(0, 3)) * 10;
             this.nReviews = parseFloat(ratingEl.getElementsByClassName("a-size-base s-underline-text")[0].textContent.replace(/[{()},.]/g, ''));
         } catch (error) {
             // No rating element found
